requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.3.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
multidict==6.0.4
openai==1.13.3
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
tqdm==4.65.0
typing_extensions==4.10.0
urllib3==1.26.6
yarl==1.9.2
